// libs/platform/list/access-data/src/lib/enums/platform-list/offers-targeting-rules.enum.ts
export enum OffersTargetingRulesIdEnum {
    ConnectionType = 1,
    MobileOperator,
    DeviceType,
    DeviceOS,
    DeviceOSVersion,
    DeviceBrand,
    DeviceModel,
    Browser,
    Language
}

export enum OffersTargetingRulesNameEnum {
    ConnectionType = 'connectionType',
    MobileOperator = 'mobileOperator',
    DeviceType = 'deviceType',
    DeviceOS = 'deviceOs',
    DeviceOSVersion = 'deviceOsVersion',
    DeviceBrand = 'deviceBrand',
    DeviceModel = 'deviceModel',
    Browser = 'browser',
    Language = 'language',
    Geo = 'geo'
}

export enum OffersTargetingRulesTranslateEnum {
    ConnectionType = 'offers_page.targeting.offers_targeting_rules.connection_type',
    MobileOperator = 'offers_page.targeting.offers_targeting_rules.mobile_operator',
    DeviceType = 'offers_page.targeting.offers_targeting_rules.device_type',
    DeviceOS = 'offers_page.targeting.offers_targeting_rules.device_os',
    DeviceOSVersion = 'offers_page.targeting.offers_targeting_rules.device_os_version',
    DeviceBrand = 'offers_page.targeting.offers_targeting_rules.device_brand',
    DeviceModel = 'offers_page.targeting.offers_targeting_rules.device_model',
    Browser = 'offers_page.targeting.offers_targeting_rules.browser',
    Language = 'offers_page.targeting.offers_targeting_rules.language'
}